requirements-handgrid.txt
torch
torchvision
datasets
Pillow
tqdm
opencv-python
onnx
onnxruntime
filterpy
pynput
